requirements.txt
flask>=2.0.0
UnityPy>=1.9.0
Pillow>=9.0.0
werkzeug>=2.0.0
